Edit: Move Avatar + CSS file into separate folder
